requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
black==26.1.0
blinker==1.9.0
cachetools==6.2.5
ccxt==4.5.34
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coincurve==21.0.0
colorlog==6.10.1
contourpy==1.3.2
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.128.0
flake8==7.3.0
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.15.0
numpy==2.2.6
optuna==4.7.0
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.4
psycopg2-binary==2.9.11
pyarrow==23.0.0
pycares==5.0.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
streamlit==1.53.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
xgboost==3.1.3
yarl==1.22.0
yfinance==1.1.0
